refactor: public/images 폴더에 icons 폴더 추가
